organizando pra funcionar com o docker compose
