+++ .env.sample
@@ -1,14 +1,18 @@
 ENVIRONMENT=development
-LOGGING_LEVL=DEBUG
+LOG_LEVEL=DEBUG
+HOST=0.0.0.0
+PORT=8080
 
 # db
 MEMORY_CONNECTION=postgresql://dialog:dialog@db:5432/dialog
 VECTORDB_CONNECTION=postgresql+psycopg://dialog:dialog@db:5432/dialog
 COLLECTION_NAME=dialog
+
 # llm and embeddings
-PARAMS_PATH=./data/params.toml
+PARAMS_PATH=/data/params/rag_with_history.toml
 OPENAI_API_KEY=????
-KNOWLEDGE_BASE_PATH=./data/knowledge_base.csv
+KNOWLEDGE_BASE_PATH=/data/knowledge_base.csv
+EMBEDDING_COLS=Document
 
 # Langsmith (Optional)
 LANGCHAIN_TRACING_V2=true
