+++ examples/knowledge_base.csv
@@ -1,0 +1,51 @@
+ArticleTitle,Question,Answer,DifficultyFromQuestioner,DifficultyFromAnswerer,ArticleFile,Document
+Abraham_Lincoln,Was Abraham Lincoln the sixteenth President of the United States?,yes,easy,easy,S08_set3_a4,"Question
+Was Abraham Lincoln the sixteenth President of the United States?
+
+Answer
+yes"
+Abraham_Lincoln,Was Abraham Lincoln the sixteenth President of the United States?,Yes.,easy,easy,S08_set3_a4,"Question
+Was Abraham Lincoln the sixteenth President of the United States?
+
+Answer
+Yes."
+Abraham_Lincoln,Did Lincoln sign the National Banking Act of 1863?,yes,easy,medium,S08_set3_a4,"Question
+Did Lincoln sign the National Banking Act of 1863?
+
+Answer
+yes"
+Abraham_Lincoln,Did Lincoln sign the National Banking Act of 1863?,Yes.,easy,easy,S08_set3_a4,"Question
+Did Lincoln sign the National Banking Act of 1863?
+
+Answer
+Yes."
+Abraham_Lincoln,Did his mother die of pneumonia?,no,easy,medium,S08_set3_a4,"Question
+Did his mother die of pneumonia?
+
+Answer
+no"
+Abraham_Lincoln,Did his mother die of pneumonia?,No.,easy,easy,S08_set3_a4,"Question
+Did his mother die of pneumonia?
+
+Answer
+No."
+Abraham_Lincoln,How many long was Lincoln's formal education?,18 months,medium,easy,S08_set3_a4,"Question
+How many long was Lincoln's formal education?
+
+Answer
+18 months"
+Abraham_Lincoln,How many long was Lincoln's formal education?,18 months.,medium,medium,S08_set3_a4,"Question
+How many long was Lincoln's formal education?
+
+Answer
+18 months."
+Abraham_Lincoln,When did Lincoln begin his political career?,1832,medium,easy,S08_set3_a4,"Question
+When did Lincoln begin his political career?
+
+Answer
+1832"
+Abraham_Lincoln,When did Lincoln begin his political career?,1832.,medium,medium,S08_set3_a4,"Question
+When did Lincoln begin his political career?
+
+Answer
+1832."
