+++ pyproject.toml
@@ -5,7 +5,6 @@ description = "Conversational Chatbot with RAG."
 authors = [
     "Luan Fernandes <luangabriel70@gmail.com>",
 ]
-readme = "README.md"
 packages = [{include = "dialog", from = "src"}]
 
 [tool.poetry.dependencies]
@@ -24,6 +23,7 @@ langchain-openai = "^0.1.0"
 pgvector = "^0.2.5"
 langchain-postgres = "^0.0.2"
 psycopg-binary = "^3.1.18"
+langchain-community = "^0.0.36"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.6.2"
