+++ src/dialog/vectorstore.py
@@ -7,6 +7,7 @@ from dialog.settings import vectordb_settings
 
 DEFAULT_DOCUMENT_PROMPT = PromptTemplate.from_template(template="{page_content}")
 
+
 def combine_documents(
     docs, document_prompt=DEFAULT_DOCUMENT_PROMPT, document_separator="\n\n"
 ):
@@ -15,7 +16,7 @@ def combine_documents(
     return document_separator.join(doc_strings)
 
 
-def get_vectorstore():
+def get_vectorstore() -> PGVector:
     return PGVector(
         collection_name=vectordb_settings.collection_name,
         connection=str(vectordb_settings.vectordb_connection),
