+++ src/scripts/start_web_server.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+uvicorn dialog.app.server:app --host $HOST --port $PORT
